2022-11-21 더미데이터, 등록화면 sql문 작성, name값 수정, VO객체 생성
