+++ YAMYAM/src/main/resources/templates/comm/pdtCodeAD.html
